refactor: remove breadcrumbs reactive var, add Breadcrumbs component to pages that use it
