+++ src/components/_App/Layout.tsx
@@ -8,7 +8,6 @@ import LeftNav from "../Navigation/LeftNav";
 import NavDrawer from "../Navigation/NavDrawer";
 import ScrollToTop from "../Navigation/ScrollToTop";
 import TopNav from "../Navigation/TopNav";
-import Breadcrumbs from "../Shared/Breadcrumbs";
 import Toast from "../Shared/Toast";
 import HeadContent from "./HeadContent";
 
@@ -35,8 +34,6 @@ const Layout = ({ children }: Props) => {
 
       <Container maxWidth="sm">
         <main role="main">
-          <Breadcrumbs />
-
           {children}
 
           <Toast />
