+++ src/pages/roles/[id]/edit.tsx
@@ -5,12 +5,12 @@ import { NextPage } from "next";
 import { useRouter } from "next/router";
 import { SyntheticEvent, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { breadcrumbsVar } from "../../../apollo/cache";
 import { useEditServerRoleQuery } from "../../../apollo/gen";
 import AddMemberTab from "../../../components/Roles/AddMemberTab";
 import DeleteRoleButton from "../../../components/Roles/DeleteRoleButton";
 import PermissionsForm from "../../../components/Roles/PermissionsForm";
 import RoleForm from "../../../components/Roles/RoleForm";
+import Breadcrumbs from "../../../components/Shared/Breadcrumbs";
 import ProgressBar from "../../../components/Shared/ProgressBar";
 import { NavigationPaths } from "../../../constants/common.constants";
 import { EditRoleTabs } from "../../../constants/role.constants";
@@ -26,28 +26,10 @@ const EditServerRole: NextPage = () => {
     variables: { id },
     skip: !id,
   });
+  const role = data?.role;
 
   const { t } = useTranslation();
   const isAboveSmall = useAboveBreakpoint("sm");
-
-  const role = data?.role;
-
-  useEffect(() => {
-    if (role) {
-      breadcrumbsVar({
-        path: asPath,
-        breadcrumbs: [
-          {
-            label: t("roles.headers.serverRoles"),
-            href: NavigationPaths.Roles,
-          },
-          {
-            label: role.name,
-          },
-        ],
-      });
-    }
-  }, [t, asPath, role]);
 
   useEffect(() => {
     if (query.tab === EditRoleTabs.Permissions) {
@@ -99,6 +81,19 @@ const EditServerRole: NextPage = () => {
 
   return (
     <>
+      <Breadcrumbs
+        path={asPath}
+        breadcrumbs={[
+          {
+            label: t("roles.headers.serverRoles"),
+            href: NavigationPaths.Roles,
+          },
+          {
+            label: role.name,
+          },
+        ]}
+      />
+
       <Card sx={{ marginBottom: 6 }}>
         <Tabs
           onChange={handleTabChange}
