+++ src/pages/users/[name]/following.tsx
@@ -7,10 +7,9 @@ import {
 import { truncate } from "lodash";
 import { NextPage } from "next";
 import { useRouter } from "next/router";
-import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { breadcrumbsVar } from "../../../apollo/cache";
 import { useFollowingQuery } from "../../../apollo/gen";
+import Breadcrumbs from "../../../components/Shared/Breadcrumbs";
 import ProgressBar from "../../../components/Shared/ProgressBar";
 import Follow from "../../../components/Users/Follow";
 import { TruncationSizes } from "../../../constants/common.constants";
@@ -36,29 +35,6 @@ const Following: NextPage = () => {
   const { t } = useTranslation();
   const isDesktop = useIsDesktop();
 
-  useEffect(() => {
-    if (user) {
-      breadcrumbsVar({
-        path: asPath,
-        breadcrumbs: [
-          {
-            label: truncate(name, {
-              length: isDesktop
-                ? TruncationSizes.Small
-                : TruncationSizes.ExtraSmall,
-            }),
-            href: getUserProfilePath(name),
-          },
-          {
-            label: t("users.labels.following", {
-              count: user.followingCount,
-            }),
-          },
-        ],
-      });
-    }
-  }, [user, t, isDesktop, asPath, name]);
-
   if (error) {
     return <Typography>{t("errors.somethingWentWrong")}</Typography>;
   }
@@ -67,23 +43,41 @@ const Following: NextPage = () => {
     return <ProgressBar />;
   }
 
-  if (!user?.following.length || !me) {
+  if (!user || !me) {
     return null;
   }
 
+  const breadcrumbs = [
+    {
+      label: truncate(name, {
+        length: isDesktop ? TruncationSizes.Small : TruncationSizes.ExtraSmall,
+      }),
+      href: getUserProfilePath(name),
+    },
+    {
+      label: t("users.labels.following", {
+        count: user.followingCount,
+      }),
+    },
+  ];
+
   return (
     <>
-      <Card>
-        <CardContent>
-          {user.following.map((followedUser) => (
-            <Follow
-              key={followedUser.id}
-              currentUserId={me.id}
-              user={followedUser}
-            />
-          ))}
-        </CardContent>
-      </Card>
+      <Breadcrumbs path={asPath} breadcrumbs={breadcrumbs} />
+
+      {!!user.followingCount && (
+        <Card>
+          <CardContent>
+            {user.following.map((followedUser) => (
+              <Follow
+                key={followedUser.id}
+                currentUserId={me.id}
+                user={followedUser}
+              />
+            ))}
+          </CardContent>
+        </Card>
+      )}
     </>
   );
 };
